Sketched some KeyGen stuff
Wrote the beginning of the keygen class and methods. It's in the client.java file for now as a quirk of development. It also comes with it's own private child class. HAven't done any encoding on the keys. And still working on it, though it may only be after graphics that I get back to it.