It's works kinda, needs more set up.
Let me explain to anyone who is actually reading this. To make it work I need to link it with bouncycastle. Now bouncy castle is a little bit of a pain to set up and I don't have the mental energy to deal with it so if you want to test it. comment out erroneus imports and on keypair generator comment out "BC" and just have "RSA"