The basics.
I commented out all the unnecessary code from Bob and Alice for now (It's still there, just commented out for now).

All this does is the basics, Alice connects to Bob on localhost thorugh port 45554 and sends a single string. This can easily be built into message and then later file.

I would really like to get bob to reply with another string and then Alice talks to the Auth server. But also that's stuff we can do in the meeting tomorrow.

Don't worry about the request Handler class for now it's just a small placeholder